Add version for TI-99/4A
This is based on the C64 version with the point-plotting routine replaced.
It requires TI Extended BASIC.
It works by placing a character on the screen and then changing the pattern for that character to display the appropriate pixels.  Since there are only 111 characters that can be redefined the total size of the image that can be displayed is limited.  I could only manage to draw the image 128 pixels wide before I ran out of available characters even though the screen is 256 pixels wide.